Fixed the expiration issue the site was having

+++ lib/tasks/deleteexpired.rake
@@ -1,7 +1,9 @@
-task :deleteexpired do 
-
-	Sentence.expired.each do |sentence| 
-		sentence.destroy 
+task :deleteexpired => :environment do 
+	Sentence.all.each do |sentence| 
+		if sentence.expired?
+			sentence.destroy 
+			sentence.save
+		end
 	end 
 	
 end
